Make `AddressSpace` a trait
This allows the user to pass any type with custom callbacks,
and not store all of ROM/RAM/etc. if they so desire.
